pyproject.toml
[project]
name = "thsr-auto-book"
version = "0.1.0"
description = "Automated THSR booking with Playwright + HTTP hybrid + CNN captcha + Discord"
requires-python = ">=3.10"
dependencies = [
    "playwright>=1.40.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "Pillow>=10.0.0",
    "opencv-python>=4.9.0",
    "numpy>=1.26.0",
    "pytesseract>=0.3.10",
    "torch>=2.0.0",
    "timm>=0.9.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["thsr_auto_book*"]